Fix many of the cross-module breaks
