+++ x/superfluid/keeper/epoch.go
@@ -109,8 +109,8 @@ func (k Keeper) UpdateOsmoEquivalentMultipliers(ctx sdk.Context, asset types.Sup
 
 		// get OSMO amount
 		bondDenom := k.sk.BondDenom(ctx)
-		osmoPoolAsset, err := pool.GetPoolAsset(bondDenom)
-		if err != nil {
+		osmoPoolAsset := pool.GetTotalLpBalances(ctx).AmountOf(bondDenom)
+		if osmoPoolAsset.IsZero() {
 			// Pool has unexpectedly removed Osmo from its assets.
 			k.Logger(ctx).Error(err.Error())
 			k.BeginUnwindSuperfluidAsset(ctx, 0, asset)
